+++ R/error_warning_note.R
@@ -1,0 +1,5 @@
+# crayon package
+
+warn <- green$bold
+note <- cyan
+error <- red$bold